requirement.txt
igraph==0.9.9
networkx==2.6.3
numpy==1.19.5
dgl ==0.7.0 + cu101
PyGSP==0.5.1
PyYAML==6.0
scikit-learn==0.24.2
scipy==1.5.4
torch==1.7.1+cu101
